+++ src/app/page.tsx
@@ -2,8 +2,6 @@ import "server-only";
 
 import clsx from "clsx";
 
-import { SearchContents } from "~/components/common/SearchContents/SearchContents";
-
 import RecentNicovideoRequestsSection from "./_components/RecentNicovideoRequestsSection";
 import RecentVideosSection from "./_components/RecentVideosSection";
 
@@ -13,21 +11,10 @@ export default async function Page() {
       className={clsx(
         ["container", "max-w-screen-2xl"],
         ["mx-auto"],
-        ["flex", "flex-col", "gap-y-8"]
+        ["flex", "flex-col", "gap-y-8"],
+        ["py-12"]
       )}
     >
-      <section
-        className={clsx(
-          [["px-2"], ["py-2"]],
-          ["rounded"],
-          ["border", "border-slate-300"]
-        )}
-      >
-        <h2 className={clsx(["text-sm"])}>検索</h2>
-        <div className={clsx(["mt-2"])}>
-          <SearchContents />
-        </div>
-      </section>
       <RecentVideosSection />
       <RecentNicovideoRequestsSection />
     </main>
